requirements.txt
asgiref==3.3.1
assertpy==1.1
attrs==20.3.0
beautifulsoup4==4.9.3
behave==1.2.6
behave-django==1.4.0
certifi==2020.12.5
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.10
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
factory-boy==3.1.0
Faker==5.0.2
idna==2.10
jsonpath==0.82
jsonref==0.2
jsonschema==3.2.0
parse==1.18.0
parse-type==0.5.2
PyJWT==1.7.1
pyrsistent==0.17.3
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.26.2
